tools/requirements.txt
pyspark
pandas
keras
tensorflow
missingno
seaborn
sklearn
tqdm
scipy
dash
gunicorn
mysql-connector
schedule
beautifulsoup4
bs4==0.0.1
certifi
chardet
idna
requests
soupsieve
urllib3
confluent-kafka
sqlalchemy